feat: implement get products
