+++ product-service/getProductsById.js
@@ -1,8 +1,11 @@
 // eslint-disable-next-line import/prefer-default-export
-import products from './data/products'
+// import products from './data/products';
+import pg_client from './data/pg_client';
+
 export async function get(event) {
   const id = event.pathParameters.productId;
-  const product = products.find(x => x.id == id);
+  const listProducts =  await pg_client.getListProducts();
+  const product = listProducts.find(x => x.id == id);
   if (product === undefined) {
     return {
       statusCode: 404,
