+++ product-service/getProductsList.js
@@ -1,8 +1,13 @@
 // eslint-disable-next-line import/prefer-default-export
 // import products from './data/products'
-import pg_client from './pg_client';
+
+import pg_client from './data/pg_client';
+
 export async function get(event) {
-  const products = await pg_client.invoke();
+
+  const products =  await pg_client.getListProducts();
+
+
   if (products === []) {
     return {
       statusCode: 404,
